+++ internal/cli/tenant/changepwd.go
@@ -88,6 +88,7 @@ func (o *ChangePwdOptions) Validate() error {
 
 func (o *ChangePwdOptions) AddFlags(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&o.Namespace, FLAG_NAMESPACE, DEFAULT_NAMESPACE, "namespace of ob tenant")
-	cmd.Flags().StringVarP(&o.Password, FLAG_PASSWD, "p", "", "new password of ob tenant")
+	cmd.Flags().StringVarP(&o.Password, FLAG_PASSWD, "p", "", "new password of ob tenant, required")
 	cmd.Flags().BoolVarP(&o.force, FLAG_FORCE, "f", DEFAULT_FORCE_FLAG, "force operation")
+	_ = cmd.MarkFlagRequired(FLAG_PASSWD)
 }